pyproject.toml
[tool.poe.tasks]
_black-check = "black --check ."
_isort-check = "isort --check-only ."
formatting-check = ["_black-check", "_isort-check"]

_black = "black ."
_isort = "isort ."
format = ["_black", "_isort"]

_pylint = "pylint loasim tests"
lint = ["_pylint"]

coverage = "pytest --cov=loasim tests"
coverage-report = "pytest --cov=loasim --cov-report=xml tests"

unittest = "pytest -vv -x tests"
e2e = "pytest -vv -x e2e"
typetest = "mypy loasim tests"

test = ["coverage", "e2e", "typetest", "lint", "formatting-check"]

[tool.black]

[tool.isort]
profile = "black"
multi_line_output = 3
skip = "third_party"
skip_gitignore = true

[tool.coverage.run]
branch = true

[tool.coverage.report]
fail_under = 85
show_missing = true
[tool.coverage.xml]
output = "coverage-reports/coverage.xml"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
incremental = false
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
ignore_missing_imports = true

[tool.poetry]
name = "loasim"
version = "0.0.0"
description = ""
authors = ["icepeng <typ0@naver.com>", "meson324 <meson324@gmail.com>"]
readme = "README.md"
repository = "https://github.com/icepeng/loasim"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
loguru = "^0.5.3"
poethepoet = "^0.10.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
isort = "^5.9.3"
coverage = "^6.0.2"
pylint = "^2.11.1"
mypy = "^0.910"
black = "^21.9b0"
pytest-cov = "^3.0.0"
requests = "^2.26.0"
syrupy = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
